Reorganize the system admin frontend folder
