+++ src/scenes/SystemAdmin/Home.jsx
@@ -1,15 +1,15 @@
-import restProvider from './rest_provider';
+import restProvider from './shared/rest_provider';
 import { fetchUtils, Admin, Resource, Layout } from 'react-admin';
 import { createBrowserHistory } from 'history';
 import React from 'react';
-import Menu from './Menu';
-import AccessCodeList from './AccessCodeList';
-import UserList from './UserList';
-import UserCreate from './UserCreate';
-import UserEdit from './UserEdit';
-import OfficeList from './OfficeList';
-import ElectronicOrderList from './ElectronicOrderList';
-import UserShow from './UserShow';
+import Menu from './shared/Menu';
+import AccessCodeList from './access_codes/AccessCodeList';
+import UserList from './office_users/UserList';
+import UserCreate from './office_users/UserCreate';
+import UserEdit from './office_users/UserEdit';
+import UserShow from './office_users/UserShow';
+import OfficeList from './offices/OfficeList';
+import ElectronicOrderList from './electronic_orders/ElectronicOrderList';
 import styles from './Home.module.scss';
 import { withContext } from 'shared/AppContext';
 import * as Cookies from 'js-cookie';
