+++ src/scenes/SystemAdmin/electronic_orders/ElectronicOrderList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { List, Datagrid, TextField } from 'react-admin';
-import AdminPagination from './AdminPagination';
-import TitleizedField from './TitleizedField';
+import AdminPagination from 'scenes/SystemAdmin/shared/AdminPagination';
+import TitleizedField from 'scenes/SystemAdmin/shared/TitleizedField';
 
 const ElectronicOrderList = props => (
   <List {...props} pagination={<AdminPagination />} perPage={25}>
