+++ src/scenes/SystemAdmin/office_users/UserList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { List, Datagrid, TextField, BooleanField } from 'react-admin';
-import AdminPagination from './AdminPagination';
+import AdminPagination from 'scenes/SystemAdmin/shared/AdminPagination';
 
 const UserList = props => (
   <List {...props} pagination={<AdminPagination />} perPage={25} bulkActionButtons={false}>
